Some little changes
