+++ .eslintrc
@@ -10,7 +10,8 @@
             "quotes": [2, "double", { "avoidEscape": true }],
             "simple-import-sort/imports": "warn",
             "simple-import-sort/exports": "warn",
-            "camelcase": "error"
+            "camelcase": "error",
+            "@typescript-eslint/indent": "off"
         }
     }]
 }
