+++ common/base_endpoint.d.ts
@@ -5,15 +5,16 @@ import winston from "winston";
 import BaseValidator from "./base_validator";
 
 declare class IBaseEndpoint {
-  validator: BaseValidator;
-  db: Knex;
-  redisClient: RedisClientType;
-  logger: winston.Logger;
-  allowNames: Array<string>;
-  model?: unknown;
+  public validator: BaseValidator;
+  public allowNames: Array<string>;
+  public model?: unknown;
 
-  constructor(db: Knex, redisClient: RedisClientType, logger: winston.Logger);
+  constructor(
+    public db: Knex,
+    public redisClient: RedisClientType,
+    public logger: winston.Logger
+  );
 
   async callEndpoint(name: string, args: RequestArgs): PromiseResponse;
-  async validate(schema: unknown, args: unknown): Promise<FunctionResponse<undefined>>;
+  async validate(schema: unknown, args: unknown): Promise<EndpointResponse<undefined>>;
 }
